requirements.txt
alembic==1.5.5
aniso8601==6.0.0
appdirs==1.4.4
astroid==2.2.5
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.4
Babel==2.9.0
bcrypt==3.2.0
cffi==1.14.5
Click==7.0
colorama==0.4.4
cryptography==3.4.6
DateTime==4.3
distlib==0.3.1
ecdsa==0.13.2
fabric==2.6.0
filelock==3.0.12
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.17.1
greenlet==1.0.0
gunicorn==20.1.0
invoke==1.5.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jose==1.0.0
jwt==1.2.0
lazy-object-proxy==1.4.0
legacy==0.1.6
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
paramiko==2.7.2
pathlib2==2.3.5
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pylint==2.3.1
PyNaCl==1.4.0
PyQt5==5.13.2
PyQt5-sip==12.8.0
pyqt5-tools==5.13.2.1.6rc1
python-dateutil==2.6.0
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2019.1
rsa==4.7.2
six==1.12.0
SQLAlchemy==1.4.8
toml==0.10.2
typed-ast==1.4.3
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.3
zope.interface==5.2.0
